Modify class as function
